ABSTRACT
‘RNSIT Placement Cell’ system is designed and developed to provide a
comprehensive platform for the organization to hire prospective
candidates with fresh ideas and rich technical skills, for the students
to register, view and apply for the suitable jobs as specified by the
organization through various facilities such as online test and
notifications.
The mentioned system is designed in accordance with the latest
technologies available for developing a fully-fledged automated web
environment to provide a rich user interface.

The Modules in this project are:
1. Administrator: Administrator has the full authority over the website.
He can view all the registered users and have the power to delete them.
He can edit the web pages and update them. He can view all the company
details also.

2. Student: A student can register himself. After registration, he will
be directed to his homepage. Here he can update his profile, change
password and see the examination details and all.

3. Company: A company can register itself, conduct online examination,
approve or disapprove candidates attending examination and provides
results about the selected candidates